requirements.txt
fastapi==0.135.1
starlette==0.52.1
uvicorn==0.41.0
httptools==0.7.1
watchfiles==1.1.1
websockets==16.0
httpx==0.28.1
httpcore==1.0.9
anyio==4.12.1
h11==0.16.0
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
asyncpg==0.31.0
greenlet==3.3.2
pandas==2.3.3
numpy==2.4.2
pyarrow==23.0.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
tzdata==2025.3
six==1.17.0
scikit-learn==1.8.0
scipy==1.17.1
joblib==1.5.3
threadpoolctl==3.6.0
sentence-transformers==5.2.3
torch==2.10.0
transformers==5.3.0
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.6.0
hf-xet==1.3.2
fsspec==2026.2.0
filelock==3.25.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.6.1
tqdm==4.67.3
regex==2026.2.28
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
blis==1.3.3
catalogue==2.0.10
confection==0.1.5
cymem==2.0.13
murmurhash==1.0.15
preshed==3.0.12
srsly==2.5.2
thinc==8.3.10
wasabi==1.1.3
weasel==0.4.3
cloudpathlib==0.23.0
smart_open==7.5.1
vaderSentiment==3.3.2
streamlit==1.55.0
plotly==6.1.2
altair==6.0.0
pydeck==0.9.1
pillow==12.1.1
tornado==6.5.4
watchdog==6.0.0
narwhals==2.18.0
blinker==1.9.0
cachetools==7.0.5
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2
python-dotenv==1.2.2
requests==2.32.5
click==8.3.1
rich==14.3.3
colorama==0.4.6
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
tenacity==9.1.4
toml==0.10.2
PyYAML==6.0.3
packaging==26.0
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
urllib3==2.6.3
wrapt==2.1.2
protobuf==6.33.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==25.4.0
Jinja2==3.1.6
MarkupSafe==3.0.3
GitPython==3.1.46
gitdb==4.0.12
smmap==5.0.3
typer==0.24.1
typer-slim==0.24.0
shellingham==1.5.4
